ldbsd.com: increase limit of image size
This addresses the 64bit boot problem of bsd4grml on Grml 2012.05-rc1.

Related to cvs revision 1.19 which states "increase the size a
grub can be to 160 KiB… the grml Schluchtenscheißer one is
ridiculously close to 128 KiB already".

Further quoting Thorsten:

| Wir haben scheints eine Größenbeschränkung auf die Imagegröße, weil
| wir immer versuchen, eine bestimmte Anzahl Bytes zu laden, da nicht
| alle im Bootloader unterstützten Dateisysteme eine Datei stat()en
| können. Dein GRUB-Image ist einfach zu fett! ;)
|
| Das Limit war mal 63½ KiB, aber das hast Du schonmal gesprengt
| gehabt… und einfach hochsetzen _kann_ auf manchen Systemen dazu
| führen, daß dann nix (auch kleinere Images) mehr geht.

Thanks: Thorsten Glaser <tg@mirbsd.org> for the fix